requirements.txt
backcall==0.2.0
cloudpickle==1.6.0
colorama==0.4.4
decorator==5.0.9
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pywin32==301
pyzmq==22.1.0
selenium==3.141.0
six==1.16.0
spyder-kernels==2.0.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
